Add support for generating the EHCont section
In the future Windows will enable Control-flow Enforcement Technology (CET aka shadow stacks). To protect the path where the context is updated during exception handling, the binary is required to enumerate valid unwind entrypoints in a dedicated section which is validated when the context is being set during exception handling.

This change includes the llvm change to generate the section that contains the appropriate symbol references in the form expected by the msvc linker.

It also includes a codegen option to enable it for rust.